feat: agregar/modificar item de carrito, eliminar item de carrito, obtener carrito de un cliente-vendedor, obtener todos los carritos de un cliente

+++ backend/ms-pedido/src/main/java/com/seminario/ms_pedido/model/DetalleCarrito.java
@@ -1,13 +1,16 @@
 package com.seminario.ms_pedido.model;
 
 import lombok.Data;
+
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 
 @Data
 @AllArgsConstructor
 public class DetalleCarrito {
     private String productoId;
-    private Double cantidad;
+    private Integer cantidad;
     private Double montoUnitario;
     private String observaciones;
 }
